Run pre-commit.

+++ hoomd/HOOMDMPI.h
@@ -566,7 +566,7 @@ class GatherTagOrder
             m_displacements[0] = 0;
             for (size_t i = 1; i < m_displacements.size(); i++)
                 {
-                m_displacements[i] = m_displacements[i-1] + m_recv_counts[i-1];
+                m_displacements[i] = m_displacements[i - 1] + m_recv_counts[i - 1];
                 }
             m_n_global_tags = m_displacements[n_ranks - 1] + m_recv_counts[n_ranks - 1];
             gathered_tags = allocateGatherBuffer<unsigned int>(m_n_global_tags);
@@ -648,7 +648,7 @@ class GatherTagOrder
             }
 
         MPI_Gatherv((void*)(local_array),
-                    local_size * sizeof(T),
+                    static_cast<int>(local_size * sizeof(T)),
                     MPI_BYTE,
                     (void*)(gathered_array),
                     m_recv_bytes.data(),
